move css to file it styles
