+++ src/dataContentPane.js
@@ -11,7 +11,7 @@
 
 import * as UI from 'solid-ui'
 import * as $rdf from 'rdflib'
-import './styles/dataContentPane.css'
+import './dataContentPane.css'
 
 const ns = UI.ns
