+++ src/imagePane.css
