+++ src/schedule/schedulePane.css
@@ -1,0 +1,29 @@
+.schedulePaneTable {
+  border-collapse: collapse;
+  background-color: white;
+  min-width: 40em;
+  min-height: 13em;
+}
+
+.schedulePaneButton {
+  background-color: #eef;
+  padding: var(--spacing-xs, 0.5em);
+  border: .5em solid white;
+  font-size: 100%;
+}
+
+.schedulePaneButtonIcon {
+  width: var(--icon-sm, 1.8em);
+  height: var(--icon-sm, 1.8em);
+}
+
+.schedulePaneNaviMenu {
+  text-align: center;
+  vertical-align: middle;
+  padding-top: 4em;
+}
+
+.schedulePaneCell {
+  padding: var(--spacing-xs, 0.3em);
+  text-align: center;
+}
