+++ src/schedule/schedulePane.js
@@ -8,7 +8,7 @@ import * as UI from 'solid-ui'
 import { authn } from 'solid-logic'
 import * as $rdf from 'rdflib'
 import formText from './formsForSchedule.ttl'
-import '../styles/schedulePane.css'
+import './schedulePane.css'
 import '../styles/utilities.css'
 
 const ns = UI.ns
