--- src/styles/humanReadablePane.css
@@ -1,8 +1,0 @@
-.humanReadableFrame {
-	border: 1px solid;
-	padding: var(--spacing-md, 1em);
-	height: var(--human-readable-height, 30em);
-	width: 800px;
-	resize: both;
-	overflow: auto;
-}
