+++ src/ui/pane.js
@@ -4,6 +4,7 @@
 
 import * as UI from 'solid-ui'
 import * as $rdf from 'rdflib'
+import '../form/formPane.css'
 const ns = UI.ns
 
 export default {
